login_register_Page/requirements.txt
flask==3.0.0
paho-mqtt==1.6.1
flask-sqlalchemy==3.1.1
werkzeug==3.0.1 